Split Python library into strata-base and strata-python packages
The single `strata` package under Tools/Python/ is now two packages:
- `strata-base` (package name: `strata`) — core DDM datatypes, Ion
  serialization, Dialect/Program classes. No Python AST dependency.
- `strata-python` (module: `strata_python`) — Python AST translation
  and the `gen` CLI. Depends on `strata-base`.

All consumers updated from `python -m strata.gen` to
`python -m strata_python.gen`. CI installs both packages separately.
