+++ StrataPython/StrataPythonTest/run_py_cbmc_tests.sh
@@ -12,7 +12,7 @@ SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 TESTS_DIR="$SCRIPT_DIR/tests"
 EXPECTED="$TESTS_DIR/cbmc_expected.txt"
 STRATA_PYTHON_DIR="$(cd "$SCRIPT_DIR/.." && pwd)"
-DIALECT="$STRATA_PYTHON_DIR/../Tools/Python/dialects/Python.dialect.st.ion"
+DIALECT="$STRATA_PYTHON_DIR/Tools/strata-python/dialects/Python.dialect.st.ion"
 PYTHON=${PYTHON:-python3}
 
 # Generate .py.ion files from .py sources (they are not committed)
@@ -20,8 +20,8 @@ echo "Generating .py.ion files from .py sources..."
 for py_file in "$TESTS_DIR"/*.py; do
   [ -f "$py_file" ] || continue
   ion_file="${py_file}.ion"
-  if ! (cd "$STRATA_PYTHON_DIR/../Tools/Python" && \
-        "$PYTHON" -m strata.gen py_to_strata --dialect "$DIALECT" "$py_file" "$ion_file") 2>/dev/null; then
+  if ! (cd "$STRATA_PYTHON_DIR/Tools/strata-python" && \
+        "$PYTHON" -m strata_python.gen py_to_strata --dialect "$DIALECT" "$py_file" "$ion_file") 2>/dev/null; then
     echo "  WARN: failed to generate $(basename "$ion_file")"
   fi
 done
