+++ StrataPython/Tools/strata-python/.gitignore
