+++ StrataPython/Tools/strata-python/pyproject.toml
@@ -1,18 +1,18 @@
 [project]
-name = "strata"
+name = "strata-python"
 version = "0.0.1"
-description = 'Python support for Strata.'
+description = "Python language support for Strata."
 requires-python = ">= 3.11"
 dependencies = [
-    "amazon.ion"
+    "strata"
 ]
 
 [tool.hatch.build.targets.wheel]
-packages = ["strata"]
+packages = ["strata_python"]
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project.scripts]
-gen = "strata.gen:main"
+strata-python-gen = "strata_python.gen:main"
