+++ Tools/Python-base/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "strata"
+version = "0.0.1"
+description = "Core Strata DDM datatypes and Ion serialization."
+requires-python = ">= 3.11"
+dependencies = [
+    "amazon.ion"
+]
+
+[tool.hatch.build.targets.wheel]
+packages = ["strata"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
