+++ Tools/Python-base/strata/base.py
